Implimenting updates to remote like to download

+++ js/discord.js
@@ -1071,7 +1071,7 @@ This code is publicly released and is restricted by its project license
                                         .catch(async (er) => {
                                             Logger.printLine("Discord", "Command was dropped, unable to get Message from Discord", "warn", er)
                                             console.error(er)
-                                            if (er && er.includes('Unknown Message')) {
+                                            if (er && er.message && er.message.includes('Unknown Message')) {
                                                 await db.query(`DELETE FROM twitter_tweets WHERE messageid = ?`, [MessageContents.messageID])
                                             }
                                             cb(true);
